% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIdata.R
\docType{data}
\name{PIdata}
\alias{PIdata}
\title{An example dataset for a PI staining experiment completed on a BioTek Cytation 5}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7570 rows and 8 columns.
}
\usage{
PIdata
}
\description{
A tibble with object level imaging data for a PI staining experiment analyzing cell cycle distribution in wildtype (WT) mouse embryonic fibroblasts (MEFs) or p53 knockout MEFs transduced with HRas-V12 oncogene connected to green fluorescent protein (GFP) by internal ribosome entry site (IRES).
}
\author{
H.G. Withers \email{hgwithers@gmail.com}
}
\keyword{datasets}
