% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpridges.R
\name{cpridges}
\alias{cpridges}
\title{Create a ridge plot with histogram and kernel density estimates for each group of data.}
\usage{
cpridges(data, signal, group)
}
\arguments{
\item{data}{A dataframe or tibble with DNA content data.}

\item{signal}{Column name containing the measurement values for DNA content staining (e.g. object integral for PI staining)}

\item{group}{Column name containing the treatment condition or unique sample ID to split the data on.}
}
\value{
A ggplot object for ridgeplots.
}
\description{
Create a ridge plot with histogram and kernel density estimates for each group of data.
}
\examples{
cpridges(data = PIdata, signal = "PI Obj Integral", group = "Cells")
}
